+++ Changes.md
@@ -1,3 +1,4 @@
+* #1182 - Add CSV for award winners of the tournament
 * #1181 - Allow team schedules to be output from the web
 * #1166 - handle Bye scores everywhere that No Show scores are handled
 * #1149 - handle changes to judging station in the schedule reports